requirements.txt
streamlit
youtube-transcript-api
fpdf
langchain
google-generativeai